api/requirements.txt
# Core
pytz==2019.3
Django==2.2.24
django-configurations==2.2
gunicorn==20.0.4
newrelic==5.4.1.134
django-guardian==2.2.0
requests==2.23.0
django-cors-headers==3.2.1

# For the persistence stores
psycopg2-binary==2.8.4
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1
django-computedfields==0.0.17
django-safedelete==0.5.5
django-expiring-token==1.0.2

# Rest apis
djangorestframework==3.11.2
Markdown==3.1.1
django-filter==2.2.0
drf-extensions==0.6.0

# Developer Tools
ipdb==0.12.3
ipython==7.11.1
mkdocs==1.0.4
flake8==3.7.9
black==19.10b0
django-extensions==2.2.8
sentry-sdk==0.18.0
requests-aws4auth==1.0.1

# Testing
mock==3.0.5
factory-boy==2.12.0
coverage==5.0.3

# Static and Media Storage
django-storages==1.8.0
boto3==1.11.1
uwsgi==2.0.18
premailer==3.7.0

# Logging
daiquiri==1.5.0
retrying==1.3.3
drf_yasg==1.17.1
